BaseBeacon gets the beacon_metadata from the object
Remove the `beacon_metadata` from the `BaseBeacon#initialize` params
list as it is available in the object. This keeps the api unchanged.

Also this commit abandons the concept of the `beacon_metadata` being a
`Hash` by default. Anyhow, this was implemented half way because the
setter had no validations. Now `beacon_metadata` can be used as its
consumers decide.

Finally, this commit extends the documentation to include more use
cases for `beacon_metadata`